El capitan

+++ zsh/install.sh
@@ -3,5 +3,7 @@
 brew install zsh
 sudo sh -c 'echo "$(which zsh)" >> /etc/shells'
 sudo chsh -s $(which zsh) $(whoami)
-sudo mv -i /etc/zshenv /etc/zprofile
+
+# older versions of OSX needed this moved
+[[ -f /etc/zshenv ]] && sudo mv -i /etc/zshenv /etc/zprofile
